feat(router-store): add provideRouterStore function (#3532)
Closes #3528
